PS-1001 Extract mapping resolution into strategies
