--- src/Writer/Table/Strategy/AllEncompassingTableStrategy.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Keboola\OutputMapping\Writer\Table\Strategy;
-
-class AllEncompassingTableStrategy extends AbstractTableStrategy
-{
-
-}
