+++ src/Writer/Table/StrategyInterface.php
@@ -3,6 +3,7 @@
 namespace Keboola\OutputMapping\Writer\Table;
 
 use Keboola\InputMapping\Staging\ProviderInterface;
+use Keboola\OutputMapping\Writer\Table\Source\SourceInterface;
 
 interface StrategyInterface
 {
@@ -15,4 +16,11 @@ interface StrategyInterface
      * @return ProviderInterface
      */
     public function getMetadataStorage();
+
+    /**
+     * @param string $sourcePath
+     * @param array $configuration
+     * @return SourceInterface[]
+     */
+    public function resolveMappings($sourcePath, array $configuration);
 }
